+++ drizzle/0080_add_thumbnail_cleaned_flag.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "media" ADD COLUMN "thumbnail_cleaned" boolean DEFAULT false NOT NULL;